fixed time savings

+++ src/Wordpress/CustomPostType/CustomPostType.php
@@ -112,6 +112,8 @@ abstract class CustomPostType
             return;
         }
 
+        $noDeleteMetaFields = ['start-time', 'end-time', 'timeframe-repetition'];
+
         /** @var Field $customField */
         foreach ($this->getCustomFields() as $customField) {
             if (current_user_can($customField->getCapability(), $post_id)) {
@@ -129,8 +131,25 @@ abstract class CustomPostType
                             $value = wpautop($value);
                         }
                         update_post_meta($post_id, $fieldName, $value);
+
+                        // Update time-fields by date-fields
+                        if(in_array($fieldName, ['start-date', 'end-date'])) {
+                            update_post_meta(
+                                $post_id,
+                                str_replace('date','time', $fieldName),
+                                date('h:i A', $value)
+                            );
+                        }
+
+                        // if we have a booking, there shall be set no repetition
+                        if($fieldName == "type" && $value == Timeframe::BOOKING_ID) {
+                            update_post_meta($post_id, 'timeframe-repetition', 'norep');
+                        }
+
                     } else {
-                        delete_post_meta($post_id, $fieldName);
+                        if(!in_array($fieldName, $noDeleteMetaFields)) {
+                            delete_post_meta($post_id, $fieldName);
+                        }
                     }
                 }
             }
